+++ packages/asc-ui/src/components/Menu/MenuInline/MenuInline.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import MenuInlineStyle, { Props } from './MenuInlineStyle'
 import MenuContext from '../MenuContext'
 import BackDrop from '../../BackDrop/BackDrop'
+import useDebounce from '../../../utils/hooks/useDebounce'
 
 const MenuInline: React.FC<Props> = ({
   children,
@@ -11,11 +12,12 @@ const MenuInline: React.FC<Props> = ({
 }) => {
   const [menuOpen, setMenuOpen] = React.useState(false)
 
-  const handleOnExpand = (expand: boolean) => {
+  // prevents the backdrop from flickering when you hover to the next menu items
+  const handleOnExpand = useDebounce((expand: boolean) => {
     setMenuOpen(expand)
 
     return onExpand && onExpand(expand)
-  }
+  }, 50)
 
   return (
     <MenuContext.Provider