+++ packages/asc-ui/src/components/SearchBarToggle/SearchBarToggle.stories.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
 import { storiesOf } from '@storybook/react'
-import { css } from '@datapunt/asc-core'
 import SearchBarToggle from './SearchBarToggle'
 
 const ControlledSearchBarToggle = () => {
@@ -11,9 +10,6 @@ const ControlledSearchBarToggle = () => {
       align="left"
       open={open}
       onOpen={setOpen}
-      css={css`
-        position: relative;
-      `}
       onWatchValue={value => {
         action(`input value changed: ${value}`)
       }}
@@ -26,14 +22,13 @@ const ControlledSearchBarToggle = () => {
 
 storiesOf('Composed/SearchBarToggle', module)
   .addDecorator(storyFn => (
-    <div style={{ padding: '40px 10px' }}>{storyFn()}</div>
+    <div style={{ padding: '40px 10px', position: 'relative' }}>
+      {storyFn()}
+    </div>
   ))
   .add('default', () => (
     <SearchBarToggle
       align="left"
-      css={css`
-        position: relative;
-      `}
       onWatchValue={value => {
         action(`input value changed: ${value}`)
       }}