+++ packages/asc-ui/src/components/SearchBarToggle/SearchBarToggleStyle.ts
@@ -1,17 +1,9 @@
 import styled from '@datapunt/asc-core'
-import {
-  themeColor,
-  showHide,
-  ShowHideTypes,
-  customCss,
-  CustomCssPropsInterface,
-} from '../../utils'
+import { themeColor, showHide, ShowHideTypes } from '../../utils'
 import SearchBarStyle from '../SearchBar/SearchBarStyle'
 import ToggleStyle from '../Toggle/ToggleStyle'
 
-export interface SearchBarToggleStyleProps
-  extends ShowHideTypes,
-    CustomCssPropsInterface {
+export interface SearchBarToggleStyleProps extends ShowHideTypes {
   open?: boolean
 }
 
@@ -31,7 +23,6 @@ const SearchBarToggleStyle = styled(ToggleStyle)<SearchBarToggleStyleProps>`
   }
 
   ${showHide()}
-  ${customCss}
 `
 
 export default SearchBarToggleStyle