+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Prefix the change with one of these keywords:
 ## Unreleased
 
 - Fixed: Update styling on `Text Button` variant
+- Changed: Keep the toggle component background colour when toggled and not toggled
 
 ## [0.17.5] - 2019-10-30