+++ CHANGELOG.md
@@ -17,7 +17,8 @@ Prefix the change with one of these keywords:
 ## Unreleased
 
 - Added: `active` prop to MenuButton to enable activeStyles (instead of only on `:hover, :focus`)
-- Changed: Keep the toggle component background colour when toggled and not toggled.
+- Fixed: Update styling on `Text Button` variant
+- Changed: Keep the toggle component background colour when toggled and not toggled
 
 ## [0.17.5] - 2019-10-30