+++ CHANGELOG.md
@@ -18,6 +18,9 @@ Prefix the change with one of these keywords:
 
 - Added: FilterTag component.
 - Changed: Keep the toggle component background colour when toggled and not toggled.
+- Added: `active` prop to MenuButton to enable activeStyles (instead of only on `:hover, :focus`)
+- Fixed: Update styling on `Text Button` variant
+- Changed: Keep the toggle component background colour when toggled and not toggled
 
 ## [0.17.5] - 2019-10-30